+++ scrapper/src/main/java/edu/java/services/jooq/JooqLinkService.java
@@ -2,6 +2,7 @@ package edu.java.services.jooq;
 
 import edu.java.api.exceptions.DoubleLinkException;
 import edu.java.api.exceptions.NoSuchLinkException;
+import edu.java.links.listener.LinkListener;
 import edu.java.model.Link;
 import edu.java.repository.jooq.JooqLinkRepository;
 import edu.java.services.LinkService;
@@ -13,13 +14,14 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class JooqLinkService implements LinkService {
     private final JooqLinkRepository jooqLinkRepository;
+    private final List<LinkListener> linkListeners;
 
     @Override
     @Transactional
     public Link add(long tgChatId, String url) {
         Link newLink = new Link();
         newLink.setCreatedAt(OffsetDateTime.now());
-        newLink.setUpdatedAt(OffsetDateTime.now());
+        newLink.setUpdatedAt(OffsetDateTime.MIN);
         newLink.setUrl(url);
 
         Link link = jooqLinkRepository.findByUrl(url).orElseGet(() -> jooqLinkRepository.addLink(newLink));
@@ -28,6 +30,8 @@ public class JooqLinkService implements LinkService {
             throw new DoubleLinkException("Link is already being tracked");
         }
 
+        linkListeners.forEach(listener -> listener.onLinkAdd(link));
+
         return link;
     }
 
@@ -35,11 +39,16 @@ public class JooqLinkService implements LinkService {
     @Transactional
     public Link remove(long tgChatId, String url) {
         Link link = jooqLinkRepository.findByUrl(url).orElseThrow(() -> new NoSuchLinkException("Link was not found"));
-        jooqLinkRepository.removeLinkByChat(tgChatId, link.getId());
+
+        if (!jooqLinkRepository.removeLinkByChat(tgChatId, link.getId())) {
+            throw new NoSuchLinkException("No such link");
+        }
 
         if (!jooqLinkRepository.findLinkInAllChats(link.getId())) {
             jooqLinkRepository.removeLink(link.getId());
         }
+
+        linkListeners.forEach(linkListener -> linkListener.onLinkRemove(link));
 
         return link;
     }
